requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Django==2.2.1
django-cors-headers==3.0.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
idna==2.8
Jinja2==2.10.1
livereload==2.6.1
Markdown==3.1.1
MarkupSafe==1.1.1
oauthlib==3.1.0
PyJWT==1.7.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
tornado==6.0.3
urllib3==1.25.6
